Added a IntentService to receive Broadcast-Intents to remotely control / configure the app.
MainActivity: Moved binding-functions to onPause() and onResume() so that the SyncThingService is only bound to the activity if the activity is active.

New class AppConfigReceiver:
Support start and stop of the SyncThingService
- restarting a running service again should not be an issue
- stop service only if "always run in background"-mode is disabled. Otherwise show a notification indicating this.

Instrumentation-tests:
- Added tests for AppConfigReceiver
- Extended MockContext to also consume stopService commands.
- testGetReadableTransferRate: Apparently the return-values have changed a bit. Adjusted the asserts to the current return-values.

SycthingService:
Added code for thread-safety in case the service still starting when it should be stopped. Then PollWebGuiAvailableTaskImpl is active and waits for the Synthing-API to become active. So that and onDestroy have to be synchronized.
Added a stopSelf() in PollWebGuiAvailableTaskImpl.onPostExecute() in case mStopScheduled was active.

Commented my change in the javadoc at onDestroy. Put a reference to that comment to .onPostExecution()