Add welcome slide for secure key generation (#4)
* Add welcome slide "key generation"

* Improve slide icon

* Add key generation via ConfigXml to welcome wizard slide
If key and config files are already present in syncthing's data folder
they won't be overwritten (as ConfigXml checks for that). It's also
no problem to go through the slides again, e.g. if the storage permission
got revoked after the first app launch granting it.

* Remove test mode

* Remove "Enjoy Syncthing."

* Improve string "Consider backing up your sync data"

* Show welcome slides only if mandatory prerequisites are
missing. Show only slides that are necessary because of
missing prerequisites. Mandatory prerequisites are
a) storage permission b) existance of keys and config
Remove key generation UI from StateDialogActivity as this
is no longer required in the main UI as we ensure generating
keys and config before launching to MainActivity.

* Minor review adjustments

* Review - Improve explanation string on config corruption