Explain maintenance status and feature request rejection in readme (#1901)
Maybe we get lucky and someone steps up. And also if not it would at least be possible to see what the state of the project is for anyone who cares to read it, and makes dealing with new feature request simpler/faster. The time I spend on keeping the app alife isn't huge, but it's still quite a few hours every now and then when some upgrade breaks things. And I would like to keep doing that, too many people are using it, but I also have zero energy to engage with feature requests (while I know tons of things that are 1000x more important that I'd address if I ever had the time/energy). And closing them with a documented reason feels better than just ignoring and accumulating tickets.