Make banner for TV track compliant (#2095)
App was rejected because of banner not filling the full page. I already
pushed these changes to the play store (via the release branch) to
iterate on play reviews. This was likely accepted, as the tv banner
rejection went away and now we are back to all file permission
rejection. "Likely" because it's not unimaginable that they are just
randomly ping-ponging between the two rejections.

Looks like our previous banner is not compliant with any standards that
ever existed. Ok maybe it did, definitely not now. Also it looks weird -
what's that background?
Use the adaptive icon guideline, created from the horizontal logo in the
main repo assets. Sizing is hopefully right, it took a bunch of trial
and error and ugly hacks to get there. Worst case the TV app launcher
thing will look off.

https://developer.android.com/design/ui/tv/guides/system/tv-app-icon-guidelines#adaptive-banner